requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.4.1
black==21.6b0
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8==3.9.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
